Ensembl Annotations

Quickly access gene annotations and transcript-to-gene (tx2gene) mappings pre-compiled from Ensembl with the biomaRt package. For gene annotables, the Entrez identifier is removed, to allow for unique Ensembl gene identifiers.

annotable(object, ...)

# S4 method for character
annotable(object, format = "gene")

Arguments

object
Object. Default usage is to provide Ensembl genome build as a string.

...
Additional arguments.

format
Desired table format, either gene, tx2gene, gene2symbol, or gene2entrez.

Value

data.frame with unique rows per gene or transcript.

Note

If the format argument is set to gene2entrez, annotable() returns a tibble with non-unique rows grouped by ensgene, instead of a data.frame.

See also

Consult the annotables package documentation (help("annotables")) for a list of currently supported genomes.

Other Gene Annotation Utilities: detectOrganism, gene2symbolFromGTF, gene2symbol, symbol2gene, tx2geneFromGTF, tx2gene

Examples

annotable("hg38") %>% glimpse

#> Observations: 63,967
#> Variables: 9
#> $ ensgene     <chr> "ENSG00000000003", "ENSG00000000005", "ENSG00000000419"...
#> $ symbol      <chr> "TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR", "...
#> $ chr         <chr> "X", "X", "20", "1", "1", "1", "1", "6", "6", "6", "1",...
#> $ start       <int> 100627109, 100584802, 50934867, 169849631, 169662007, 2...
#> $ end         <int> 100639991, 100599885, 50958555, 169894267, 169854080, 2...
#> $ strand      <int> -1, 1, -1, -1, 1, -1, 1, -1, -1, 1, -1, 1, -1, 1, 1, 1,...
#> $ biotype     <chr> "protein_coding", "protein_coding", "protein_coding", "...
#> $ description <chr> "tetraspanin 6 [Source:HGNC Symbol;Acc:HGNC:11858]", "t...
#> $ broadClass  <chr> "coding", "coding", "coding", "coding", "coding", "codi...