Ensembl Annotations

Quickly access gene annotations and transcript-to-gene (tx2gene) mappings pre-compiled from Ensembl with the biomaRt package. For gene annotables, the Entrez identifier is removed, to allow for unique Ensembl gene identifiers.

annotable(object)

gene2entrez(object)

gene2symbol(object)

tx2gene(object)

# S4 method for character
annotable(object)

# S4 method for character
gene2entrez(object)

# S4 method for character
gene2symbol(object)

# S4 method for character
tx2gene(object)

Arguments

object
Object. Default usage is to provide Ensembl genome build as a string.

Value

data.frame with unique rows per gene or transcript.

Note

If the format argument is set to gene2entrez, annotable() returns a tibble with non-unique rows grouped by ensgene, instead of a data.frame.

See also

Consult the annotables package documentation (help("annotables")) for a list of currently supported genomes.

Examples

annotable("hg38") %>% glimpse

#> Observations: 63,898
#> Variables: 9
#> $ ensgene     <chr> "ENSG00000000003", "ENSG00000000005", "ENSG00000000419"...
#> $ symbol      <chr> "TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR", "...
#> $ chr         <chr> "X", "X", "20", "1", "1", "1", "1", "6", "6", "6", "1",...
#> $ start       <int> 100627109, 100584802, 50934867, 169849631, 169662007, 2...
#> $ end         <int> 100639991, 100599885, 50958555, 169894267, 169854080, 2...
#> $ strand      <int> -1, 1, -1, -1, 1, -1, 1, -1, -1, 1, -1, 1, -1, 1, 1, 1,...
#> $ biotype     <chr> "protein_coding", "protein_coding", "protein_coding", "...
#> $ description <chr> "tetraspanin 6 [Source:HGNC Symbol;Acc:HGNC:11858]", "t...
#> $ broad_class <chr> "coding", "coding", "coding", "coding", "coding", "codi...

detectOrganism("ENSMUSG00000000001") %>% gene2symbol %>% head

#>                               ensgene symbol
#> ENSMUSG00000000001 ENSMUSG00000000001  Gnai3
#> ENSMUSG00000000003 ENSMUSG00000000003   Pbsn
#> ENSMUSG00000000028 ENSMUSG00000000028  Cdc45
#> ENSMUSG00000000031 ENSMUSG00000000031    H19
#> ENSMUSG00000000037 ENSMUSG00000000037  Scml2
#> ENSMUSG00000000049 ENSMUSG00000000049   Apoh