Summarize columns with toString().

Collapse column data into a single row summary columnwize using toString().

toStringSummarize(data)

toStringSummarise(data)

Arguments

data
Data with rows and columns (e.g. data frame, matrix)

Value

Summarized data frame that has been collapsed to a single toString()-formatted row, separated by commas.

Examples

toStringSummarize(head(starwars))

#>                                                                name
#> 1 Luke Skywalker, C-3PO, R2-D2, Darth Vader, Leia Organa, Owen Lars
#>                        height                     mass
#> 1 172, 167, 96, 202, 150, 178 77, 75, 32, 136, 49, 120
#>                        hair_color                            skin_color
#> 1 blond, none, brown, brown, grey fair, gold, white, blue, white, light
#>                  eye_color            birth_year       gender
#> 1 blue, yellow, red, brown 19, 112, 33, 41.9, 52 male, female
#>                   homeworld      species
#> 1 Tatooine, Naboo, Alderaan Human, Droid
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      films
#> 1 c("Revenge of the Sith", "Return of the Jedi", "The Empire Strikes Back", "A New Hope", "The Force Awakens"), c("Attack of the Clones", "The Phantom Menace", "Revenge of the Sith", "Return of the Jedi", "The Empire Strikes Back", "A New Hope"), c("Attack of the Clones", "The Phantom Menace", "Revenge of the Sith", "Return of the Jedi", "The Empire Strikes Back", "A New Hope", "The Force Awakens"), c("Revenge of the Sith", "Return of the Jedi", "The Empire Strikes Back", "A New Hope"), c("Attack of the Clones", "Revenge of the Sith", "A New Hope")
#>                                                                         vehicles
#> 1 c("Snowspeeder", "Imperial Speeder Bike"), character(0), Imperial Speeder Bike
#>                                                        starships
#> 1 c("X-wing", "Imperial shuttle"), character(0), TIE Advanced x1