Ensembl Annotations

Quickly access gene annotations and transcript-to-gene (tx2gene) mappings pre-compiled from Ensembl with the biomaRt package. For gene annotables, the Entrez identifier is removed, to allow for unique Ensembl gene identifiers.

annotable(object, ...)

# S4 method for character
annotable(object, format = "gene", release = 88L)

Arguments

object
Object. Default usage is to provide Ensembl genome build as a string.

...
Additional arguments (for the S4 generic definition).

format
Desired table format, either gene, tx2gene, gene2symbol, or gene2entrez.

release
Ensembl release version. This function defaults to using the most recent annotations available on AnnotationHub.

Value

data.frame with unique rows per gene or transcript.

Note

If the format argument is set to gene2entrez, annotable() returns a tibble with non-unique rows grouped by ensgene, instead of a data.frame.

See also

Consult the annotables package documentation (help("annotables")) for a list of currently supported genomes.

Other Gene Annotation Utilities: detectOrganism, gene2symbolFromGTF, gene2symbol, symbol2gene, tx2geneFromGTF, tx2gene

Examples

annotable("Mus musculus") %>% glimpse

#> Mus musculus Ensembl 88 annotations

#> snapshotDate(): 2017-04-25

#> loading from cache '/Users/mike//.AnnotationHub/60464'

#> Observations: 51,158
#> Variables: 5
#> $ ensgene     <chr> "ENSMUSG00000000001", "ENSMUSG00000000003", "ENSMUSG000...
#> $ symbol      <chr> "Gnai3", "Pbsn", "Cdc45", "H19", "Scml2", "Apoh", "Narf...
#> $ description <chr> "guanine nucleotide binding protein (G protein), alpha ...
#> $ biotype     <chr> "protein_coding", "protein_coding", "protein_coding", "...
#> $ broadClass  <chr> "coding", "coding", "coding", "noncoding", "coding", "c...